fix: delete service functionality in separate folder
