+++ packages/odata-service-writer/src/updates.ts
@@ -1,92 +1,12 @@
 import { render } from 'ejs';
 import type { Editor } from 'mem-fs-editor';
-import { dirname, join, normalize, posix } from 'path';
+import { join, normalize, posix } from 'path';
 import { t } from './i18n';
 import type { OdataService, CdsAnnotationsInfo, EdmxAnnotationsInfo } from './types';
 import semVer from 'semver';
 import prettifyXml from 'prettify-xml';
-import type { ManifestNamespace } from '@sap-ux/project-access';
-import { getMinimumUI5Version, type Manifest, hasUI5CliV3 } from '@sap-ux/project-access';
-
-/**
- * Internal function that deletes EDMX file for dataSource.
- *
- * @param fs - the memfs editor instance
- * @param manifestPath - the root path of an existing UI5 application
- * @param dataSource - name of the OData service instance
- */
-async function deleteEDMXFileForDataSource(
-    fs: Editor,
-    manifestPath: string,
-    dataSource: ManifestNamespace.DataSource
-): Promise<void> {
-    const serviceSettings = dataSource.settings || {};
-    if (serviceSettings?.localUri) {
-        const localUriPath = join(dirname(manifestPath), serviceSettings?.localUri);
-        if (fs.exists(localUriPath)) {
-            // delete the local data source file
-            await fs.delete(localUriPath);
-        }
-    }
-}
-
-/**
- * Internal function that deletes service from the manifest.json based on the given service name.
- *
- * @param basePath - the root path of an existing UI5 application
- * @param serviceName - name of the OData service instance
- * @param fs - the memfs editor instance
- */
-export async function deleteServiceFromManifest(basePath: string, serviceName: string, fs: Editor): Promise<void> {
-    const manifestPath = join(basePath, 'webapp', 'manifest.json');
-    // Get component app id
-    const manifest = fs.readJSON(manifestPath) as unknown as Manifest;
-    const appProp = 'sap.app';
-    const appid = manifest?.[appProp]?.id;
-    // Throw if required property is not found manifest.json
-    if (!appid) {
-        throw new Error(
-            t('error.requiredProjectPropertyNotFound', { property: `'${appProp}'.id`, path: manifestPath })
-        );
-    }
-    const dataSources = manifest?.[appProp]?.dataSources;
-    if (dataSources?.[serviceName]) {
-        await deleteEDMXFileForDataSource(fs, manifestPath, dataSources?.[serviceName]);
-    }
-    const serviceSettings = dataSources?.[serviceName]?.settings;
-
-    // Check for linked backend annotations and delete if found.
-    if (serviceSettings?.annotations && serviceSettings.annotations.length > 0) {
-        for (const datasourceKey of serviceSettings.annotations) {
-            const annotationDatasource = dataSources?.[datasourceKey];
-            if (annotationDatasource?.type === 'ODataAnnotation') {
-                if (annotationDatasource.uri === annotationDatasource?.settings?.localUri) {
-                    // This is localAnnotaton file. Do not delete it.
-                } else if (annotationDatasource) {
-                    await deleteEDMXFileForDataSource(fs, manifestPath, annotationDatasource);
-                    // delete dataSource from manifest
-                    delete dataSources?.[datasourceKey];
-                }
-            }
-        }
-    }
-    // delete dataSource from manifest
-    if (dataSources?.[serviceName]) {
-        delete dataSources[serviceName];
-    }
-    const modelsProp = 'sap.ui5';
-    // delete models for this service
-    const models = manifest?.[modelsProp]?.models;
-    if (models) {
-        for (const modelKey of Object.keys(models)) {
-            const modelObj = models[modelKey];
-            if (modelObj?.dataSource === serviceName) {
-                delete models[modelKey];
-            }
-        }
-    }
-    fs.writeJSON(manifestPath, manifest);
-}
+import type { Manifest } from '@sap-ux/project-access';
+import { getMinimumUI5Version, hasUI5CliV3 } from '@sap-ux/project-access';
 
 /**
  * Internal function that updates the manifest.json based on the given service configuration.
@@ -223,7 +143,7 @@ export async function updateCdsFilesWithAnnotations(
  * @param minUI5Version - The minimum UI5 version.
  * @returns updated model settings.
  */
-function getModelSettings(minUI5Version: string | undefined) {
+function getModelSettings(minUI5Version: string | undefined): { includeSynchronizationMode: boolean } {
     let includeSynchronizationMode = false;
     if (minUI5Version) {
         includeSynchronizationMode = semVer.satisfies(minUI5Version, '<=1.110');
@@ -237,7 +157,7 @@ function getModelSettings(minUI5Version: string | undefined) {
  * @param fs - the memfs editor instance
  * @param addMockServer true if the mocksever middleware needs to be added as well
  */
-export function updatePackageJson(path: string, fs: Editor, addMockServer: boolean) {
+export function updatePackageJson(path: string, fs: Editor, addMockServer: boolean): void {
     const packageJson = JSON.parse(fs.read(path));
     packageJson.devDependencies = packageJson.devDependencies ?? {};
     if (!hasUI5CliV3(packageJson.devDependencies)) {
